Implement type promotion and weak types (#8)
* Add currently failing tests for weak type behavior

* Add more weak type tests

* Implement type promotion and weak types

* Fix bug in loading weights for clip inference

* Remove scalar() function

* Clean up some manual type usage

* Fix minor check error

+++ packages/optax/src/transform.ts
@@ -47,12 +47,12 @@ export function scaleByAdam({
       };
       mu = treeUpdateMoment(tree.ref(updates), mu, b1, 1);
       nu = treeUpdateMoment(tree.ref(updates), nu, b2, 2);
-      count = count.add(u32(1));
+      count = count.add(1);
       let muHat: typeof mu;
       if (nesterov) {
         muHat = tree.map(
           (m: np.Array, g: np.Array) => m.mul(b1).add(g.mul(1 - b1)),
-          treeBiasCorrection(tree.ref(mu), b1, count.ref.add(u32(1))),
+          treeBiasCorrection(tree.ref(mu), b1, count.ref.add(1)),
           treeBiasCorrection(tree.ref(updates), b1, count.ref),
         );
       } else {
